:recycle: :bug: refactor `proctree` package to fix errors not propagating to `WaitReady()` properly
